+++ mmd_uuunyaa_tools/editors/armatures/rigify.py
@@ -10,7 +10,6 @@ from typing import Callable, Dict, Iterable
 import bpy
 from mathutils import Color, Euler, Matrix, Vector
 from mmd_uuunyaa_tools.editors.armatures import (PATH_BLENDS_RIGSHAPELIBRARY,
-                                                 ArmatureObjectABC,
                                                  ControlType, DataPath,
                                                  GroupType, MMDArmatureObject,
                                                  MMDBindInfo, MMDBindType,
@@ -151,237 +150,6 @@ def _create_binders() -> Dict[MMDBindType, Callable]:
 
 
 binders: Dict[MMDBindType, Callable] = _create_binders()
-
-
-class MetarigArmatureObject(ArmatureObjectABC):
-    def fit_scale(self, mmd_armature_object: MMDArmatureObject):
-        rigify_height = self.bones['spine.004'].head_local[2]
-        mmd_height = mmd_armature_object.strict_bones['首'].head_local[2]
-
-        scale = mmd_height / rigify_height
-        self.raw_object.scale = (scale, scale, scale)
-        bpy.ops.object.transform_apply(scale=True)
-
-    def fit_bones(self, mmd_armature_object: MMDArmatureObject):
-        # pylint: disable=too-many-statements
-        metarig_edit_bones = self.edit_bones
-        mmd_edit_bones = mmd_armature_object.strict_edit_bones
-        metarig_edit_bones['spine.001'].head = mmd_edit_bones['上半身'].head
-
-        if MMDBoneType.UPPER_BODY_2 in mmd_armature_object.exist_bone_types:
-            metarig_edit_bones['spine.002'].head = mmd_edit_bones['上半身2'].head
-            metarig_edit_bones['spine.003'].head = self.to_bone_center(mmd_edit_bones['上半身2'])
-            metarig_edit_bones['spine.003'].tail = mmd_edit_bones['上半身2'].tail
-        else:
-            metarig_edit_bones['spine.002'].head = self.to_bone_center(mmd_edit_bones['上半身'])
-            metarig_edit_bones['spine.003'].head = self.to_bone_stretch(mmd_edit_bones['上半身'], 3/4.000)
-            metarig_edit_bones['spine.003'].tail = mmd_edit_bones['上半身'].tail
-
-        metarig_edit_bones['spine.004'].head = mmd_edit_bones['首'].head
-        metarig_edit_bones['spine.004'].tail = self.to_bone_center(mmd_edit_bones['首'])
-        metarig_edit_bones['spine.005'].tail = mmd_edit_bones['首'].tail
-        metarig_edit_bones['spine.006'].tail = mmd_edit_bones['頭'].tail
-
-        metarig_edit_bones['face'].head = mmd_edit_bones['頭'].head
-        metarig_edit_bones['face'].tail = self.to_bone_center(mmd_edit_bones['頭'])
-
-        metarig_edit_bones['shoulder.L'].head = mmd_edit_bones['左肩'].head
-        metarig_edit_bones['shoulder.L'].tail = mmd_edit_bones['左肩'].tail
-        metarig_edit_bones['shoulder.R'].head = mmd_edit_bones['右肩'].head
-        metarig_edit_bones['shoulder.R'].tail = mmd_edit_bones['右肩'].tail
-        metarig_edit_bones['upper_arm.L'].head = mmd_edit_bones['左腕'].head
-        metarig_edit_bones['upper_arm.R'].head = mmd_edit_bones['右腕'].head
-        metarig_edit_bones['forearm.L'].head = mmd_edit_bones['左ひじ'].head
-        metarig_edit_bones['forearm.L'].tail = mmd_edit_bones['左ひじ'].tail
-        metarig_edit_bones['forearm.R'].head = mmd_edit_bones['右ひじ'].head
-        metarig_edit_bones['forearm.R'].tail = mmd_edit_bones['右ひじ'].tail
-
-        metarig_edit_bones['hand.L'].tail = mmd_edit_bones['左手首'].tail
-        metarig_edit_bones['hand.R'].tail = mmd_edit_bones['右手首'].tail
-
-        if MMDBoneType.THUMB_0 in mmd_armature_object.exist_bone_types:
-            metarig_edit_bones['thumb.01.L'].head = mmd_edit_bones['左親指０'].head
-            metarig_edit_bones['thumb.01.L'].tail = mmd_edit_bones['左親指０'].tail
-            metarig_edit_bones['thumb.01.R'].head = mmd_edit_bones['右親指０'].head
-            metarig_edit_bones['thumb.01.R'].tail = mmd_edit_bones['右親指０'].tail
-        else:
-            metarig_edit_bones['thumb.01.L'].head = mmd_edit_bones['左親指１'].head - mmd_edit_bones['左親指１'].vector
-            metarig_edit_bones['thumb.01.L'].tail = mmd_edit_bones['左親指１'].head
-            metarig_edit_bones['thumb.01.R'].head = mmd_edit_bones['右親指１'].head - mmd_edit_bones['右親指１'].vector
-            metarig_edit_bones['thumb.01.R'].tail = mmd_edit_bones['右親指１'].head
-
-        metarig_edit_bones['thumb.02.L'].tail = mmd_edit_bones['左親指１'].tail
-        metarig_edit_bones['thumb.02.R'].tail = mmd_edit_bones['右親指１'].tail
-        metarig_edit_bones['thumb.03.L'].tail = mmd_edit_bones['左親指２'].tail
-        metarig_edit_bones['thumb.03.R'].tail = mmd_edit_bones['右親指２'].tail
-        metarig_edit_bones['palm.01.L'].head = self.to_center(mmd_edit_bones['左人指１'].head, mmd_edit_bones['左ひじ'].tail)
-        metarig_edit_bones['palm.01.L'].tail = mmd_edit_bones['左人指１'].head
-        metarig_edit_bones['palm.01.R'].head = self.to_center(mmd_edit_bones['右人指１'].head, mmd_edit_bones['右ひじ'].tail)
-        metarig_edit_bones['palm.01.R'].tail = mmd_edit_bones['右人指１'].head
-        metarig_edit_bones['f_index.01.L'].head = mmd_edit_bones['左人指１'].head
-        metarig_edit_bones['f_index.01.L'].tail = mmd_edit_bones['左人指１'].tail
-        metarig_edit_bones['f_index.01.R'].head = mmd_edit_bones['右人指１'].head
-        metarig_edit_bones['f_index.01.R'].tail = mmd_edit_bones['右人指１'].tail
-        metarig_edit_bones['f_index.02.L'].tail = mmd_edit_bones['左人指２'].tail
-        metarig_edit_bones['f_index.02.R'].tail = mmd_edit_bones['右人指２'].tail
-        metarig_edit_bones['f_index.03.L'].tail = mmd_edit_bones['左人指３'].tail
-        metarig_edit_bones['f_index.03.R'].tail = mmd_edit_bones['右人指３'].tail
-        metarig_edit_bones['palm.02.L'].head = self.to_center(mmd_edit_bones['左中指１'].head, mmd_edit_bones['左ひじ'].tail)
-        metarig_edit_bones['palm.02.L'].tail = mmd_edit_bones['左中指１'].head
-        metarig_edit_bones['palm.02.R'].head = self.to_center(mmd_edit_bones['右中指１'].head, mmd_edit_bones['右ひじ'].tail)
-        metarig_edit_bones['palm.02.R'].tail = mmd_edit_bones['右中指１'].head
-        metarig_edit_bones['f_middle.01.L'].head = mmd_edit_bones['左中指１'].head
-        metarig_edit_bones['f_middle.01.L'].tail = mmd_edit_bones['左中指１'].tail
-        metarig_edit_bones['f_middle.01.R'].head = mmd_edit_bones['右中指１'].head
-        metarig_edit_bones['f_middle.01.R'].tail = mmd_edit_bones['右中指１'].tail
-        metarig_edit_bones['f_middle.02.L'].tail = mmd_edit_bones['左中指２'].tail
-        metarig_edit_bones['f_middle.02.R'].tail = mmd_edit_bones['右中指２'].tail
-        metarig_edit_bones['f_middle.03.L'].tail = mmd_edit_bones['左中指３'].tail
-        metarig_edit_bones['f_middle.03.R'].tail = mmd_edit_bones['右中指３'].tail
-        metarig_edit_bones['palm.03.L'].head = self.to_center(mmd_edit_bones['左薬指１'].head, mmd_edit_bones['左ひじ'].tail)
-        metarig_edit_bones['palm.03.L'].tail = mmd_edit_bones['左薬指１'].head
-        metarig_edit_bones['palm.03.R'].head = self.to_center(mmd_edit_bones['右薬指１'].head, mmd_edit_bones['右ひじ'].tail)
-        metarig_edit_bones['palm.03.R'].tail = mmd_edit_bones['右薬指１'].head
-        metarig_edit_bones['f_ring.01.L'].head = mmd_edit_bones['左薬指１'].head
-        metarig_edit_bones['f_ring.01.L'].tail = mmd_edit_bones['左薬指１'].tail
-        metarig_edit_bones['f_ring.01.R'].head = mmd_edit_bones['右薬指１'].head
-        metarig_edit_bones['f_ring.01.R'].tail = mmd_edit_bones['右薬指１'].tail
-        metarig_edit_bones['f_ring.02.L'].tail = mmd_edit_bones['左薬指２'].tail
-        metarig_edit_bones['f_ring.02.R'].tail = mmd_edit_bones['右薬指２'].tail
-        metarig_edit_bones['f_ring.03.L'].tail = mmd_edit_bones['左薬指３'].tail
-        metarig_edit_bones['f_ring.03.R'].tail = mmd_edit_bones['右薬指３'].tail
-        metarig_edit_bones['palm.04.L'].head = self.to_center(mmd_edit_bones['左小指１'].head, mmd_edit_bones['左ひじ'].tail)
-        metarig_edit_bones['palm.04.L'].tail = mmd_edit_bones['左小指１'].head
-        metarig_edit_bones['palm.04.R'].head = self.to_center(mmd_edit_bones['右小指１'].head, mmd_edit_bones['右ひじ'].tail)
-        metarig_edit_bones['palm.04.R'].tail = mmd_edit_bones['右小指１'].head
-        metarig_edit_bones['f_pinky.01.L'].head = mmd_edit_bones['左小指１'].head
-        metarig_edit_bones['f_pinky.01.L'].tail = mmd_edit_bones['左小指１'].tail
-        metarig_edit_bones['f_pinky.01.R'].head = mmd_edit_bones['右小指１'].head
-        metarig_edit_bones['f_pinky.01.R'].tail = mmd_edit_bones['右小指１'].tail
-        metarig_edit_bones['f_pinky.02.L'].tail = mmd_edit_bones['左小指２'].tail
-        metarig_edit_bones['f_pinky.02.R'].tail = mmd_edit_bones['右小指２'].tail
-        metarig_edit_bones['f_pinky.03.L'].tail = mmd_edit_bones['左小指３'].tail
-        metarig_edit_bones['f_pinky.03.R'].tail = mmd_edit_bones['右小指３'].tail
-
-        metarig_edit_bones['spine'].head = mmd_edit_bones['下半身'].tail
-        metarig_edit_bones['spine'].tail = mmd_edit_bones['下半身'].head
-
-        metarig_edit_bones['pelvis.L'].head = mmd_edit_bones['下半身'].tail
-        metarig_edit_bones['pelvis.R'].head = mmd_edit_bones['下半身'].tail
-        metarig_edit_bones['pelvis.L'].tail[1:3] = [mmd_edit_bones['下半身'].tail[1]-metarig_edit_bones['spine'].length/2, mmd_edit_bones['下半身'].head[2]]
-        metarig_edit_bones['pelvis.R'].tail[1:3] = [mmd_edit_bones['下半身'].tail[1]-metarig_edit_bones['spine'].length/2, mmd_edit_bones['下半身'].head[2]]
-
-        metarig_edit_bones['thigh.L'].head = mmd_edit_bones['左足'].head
-        metarig_edit_bones['thigh.L'].tail = mmd_edit_bones['左足'].tail
-        metarig_edit_bones['thigh.R'].head = mmd_edit_bones['右足'].head
-        metarig_edit_bones['thigh.R'].tail = mmd_edit_bones['右足'].tail
-        metarig_edit_bones['shin.L'].tail = mmd_edit_bones['左ひざ'].tail
-        metarig_edit_bones['shin.R'].tail = mmd_edit_bones['右ひざ'].tail
-        metarig_edit_bones['foot.L'].tail = mmd_edit_bones['左足首'].tail
-        metarig_edit_bones['foot.R'].tail = mmd_edit_bones['右足首'].tail
-        metarig_edit_bones['heel.02.L'].tail = mmd_edit_bones['左ひざ'].tail.copy()
-        metarig_edit_bones['heel.02.L'].tail[0] += +mmd_edit_bones['左ひざ'].length / 8
-        metarig_edit_bones['heel.02.L'].tail[1] += +mmd_edit_bones['左ひざ'].length / 10
-        metarig_edit_bones['heel.02.L'].tail[2] = 0
-        metarig_edit_bones['heel.02.L'].head = mmd_edit_bones['左ひざ'].tail.copy()
-        metarig_edit_bones['heel.02.L'].head[0] += -mmd_edit_bones['左ひざ'].length / 8
-        metarig_edit_bones['heel.02.L'].head[1] += +mmd_edit_bones['左ひざ'].length / 10
-        metarig_edit_bones['heel.02.L'].head[2] = 0
-        metarig_edit_bones['heel.02.R'].tail = mmd_edit_bones['右ひざ'].tail.copy()
-        metarig_edit_bones['heel.02.R'].tail[0] += -mmd_edit_bones['右ひざ'].length / 8
-        metarig_edit_bones['heel.02.R'].tail[1] += +mmd_edit_bones['右ひざ'].length / 10
-        metarig_edit_bones['heel.02.R'].tail[2] = 0
-        metarig_edit_bones['heel.02.R'].head = mmd_edit_bones['右ひざ'].tail.copy()
-        metarig_edit_bones['heel.02.R'].head[0] += +mmd_edit_bones['右ひざ'].length / 8
-        metarig_edit_bones['heel.02.R'].head[1] += +mmd_edit_bones['右ひざ'].length / 10
-        metarig_edit_bones['heel.02.R'].head[2] = 0
-        metarig_edit_bones['toe.L'].tail = mmd_edit_bones['左足首'].tail + Vector([+0.0, -mmd_edit_bones['左足首'].length / 2, +0.0])
-        metarig_edit_bones['toe.R'].tail = mmd_edit_bones['右足首'].tail + Vector([+0.0, -mmd_edit_bones['右足首'].length / 2, +0.0])
-
-        # fix straight finger bend problem
-        # https://blenderartists.org/t/rigify-fingers-issue/1218987
-        # limbs.super_finger
-        metarig_edit_bones['thumb.01.L'].roll += math.radians(-45)
-        metarig_edit_bones['thumb.02.L'].roll += math.radians(-45)
-        metarig_edit_bones['thumb.03.L'].roll += math.radians(-45)
-        metarig_edit_bones['f_index.01.L'].roll += math.radians(-45)
-        metarig_edit_bones['f_index.02.L'].roll += math.radians(-45)
-        metarig_edit_bones['f_index.03.L'].roll += math.radians(-45)
-        metarig_edit_bones['f_middle.01.L'].roll += math.radians(-45)
-        metarig_edit_bones['f_middle.02.L'].roll += math.radians(-45)
-        metarig_edit_bones['f_middle.03.L'].roll += math.radians(-45)
-        metarig_edit_bones['f_ring.01.L'].roll += math.radians(-45)
-        metarig_edit_bones['f_ring.02.L'].roll += math.radians(-45)
-        metarig_edit_bones['f_ring.03.L'].roll += math.radians(-45)
-        metarig_edit_bones['f_pinky.01.L'].roll += math.radians(-45)
-        metarig_edit_bones['f_pinky.02.L'].roll += math.radians(-45)
-        metarig_edit_bones['f_pinky.03.L'].roll += math.radians(-45)
-
-        metarig_edit_bones['thumb.01.R'].roll += math.radians(+45)
-        metarig_edit_bones['thumb.02.R'].roll += math.radians(+45)
-        metarig_edit_bones['thumb.03.R'].roll += math.radians(+45)
-        metarig_edit_bones['f_index.01.R'].roll += math.radians(+45)
-        metarig_edit_bones['f_index.02.R'].roll += math.radians(+45)
-        metarig_edit_bones['f_index.03.R'].roll += math.radians(+45)
-        metarig_edit_bones['f_middle.01.R'].roll += math.radians(+45)
-        metarig_edit_bones['f_middle.02.R'].roll += math.radians(+45)
-        metarig_edit_bones['f_middle.03.R'].roll += math.radians(+45)
-        metarig_edit_bones['f_ring.01.R'].roll += math.radians(+45)
-        metarig_edit_bones['f_ring.02.R'].roll += math.radians(+45)
-        metarig_edit_bones['f_ring.03.R'].roll += math.radians(+45)
-        metarig_edit_bones['f_pinky.01.R'].roll += math.radians(+45)
-        metarig_edit_bones['f_pinky.02.R'].roll += math.radians(+45)
-        metarig_edit_bones['f_pinky.03.R'].roll += math.radians(+45)
-
-        # fix elbow pole problem
-        # https://blenderartists.org/t/rigify-elbow-problem/565285
-        metarig_edit_bones['upper_arm.L'].tail += Vector([0, +0.001, 0])
-        metarig_edit_bones['upper_arm.R'].tail += Vector([0, +0.001, 0])
-
-        # remove unused mmd_edit_bones
-        remove_metarig_edit_bones = [
-            metarig_edit_bones['breast.L'],
-            metarig_edit_bones['breast.R'],
-        ]
-        for metarig_bone in remove_metarig_edit_bones:
-            metarig_edit_bones.remove(metarig_bone)
-
-    def set_rigify_parameters(self):
-        metarig_pose_bones = self.pose_bones
-
-        # fix straight finger bend problem
-        # https://blenderartists.org/t/rigify-fingers-issue/1218987
-        # limbs.super_finger
-        metarig_pose_bones['thumb.01.L'].rigify_parameters.primary_rotation_axis = 'X'
-        metarig_pose_bones['thumb.01.L'].rigify_parameters.make_extra_ik_control = True
-        metarig_pose_bones['thumb.01.R'].rigify_parameters.primary_rotation_axis = 'X'
-        metarig_pose_bones['thumb.01.R'].rigify_parameters.make_extra_ik_control = True
-        metarig_pose_bones['f_index.01.L'].rigify_parameters.primary_rotation_axis = 'X'
-        metarig_pose_bones['f_index.01.L'].rigify_parameters.make_extra_ik_control = True
-        metarig_pose_bones['f_index.01.R'].rigify_parameters.primary_rotation_axis = 'X'
-        metarig_pose_bones['f_index.01.R'].rigify_parameters.make_extra_ik_control = True
-        metarig_pose_bones['f_middle.01.L'].rigify_parameters.primary_rotation_axis = 'X'
-        metarig_pose_bones['f_middle.01.L'].rigify_parameters.make_extra_ik_control = True
-        metarig_pose_bones['f_middle.01.R'].rigify_parameters.primary_rotation_axis = 'X'
-        metarig_pose_bones['f_middle.01.R'].rigify_parameters.make_extra_ik_control = True
-        metarig_pose_bones['f_ring.01.L'].rigify_parameters.primary_rotation_axis = 'X'
-        metarig_pose_bones['f_ring.01.L'].rigify_parameters.make_extra_ik_control = True
-        metarig_pose_bones['f_ring.01.R'].rigify_parameters.primary_rotation_axis = 'X'
-        metarig_pose_bones['f_ring.01.R'].rigify_parameters.make_extra_ik_control = True
-        metarig_pose_bones['f_pinky.01.L'].rigify_parameters.primary_rotation_axis = 'X'
-        metarig_pose_bones['f_pinky.01.L'].rigify_parameters.make_extra_ik_control = True
-        metarig_pose_bones['f_pinky.01.R'].rigify_parameters.primary_rotation_axis = 'X'
-        metarig_pose_bones['f_pinky.01.R'].rigify_parameters.make_extra_ik_control = True
-
-        # fix straight arm IK problem
-        # limbs.super_limb
-        metarig_pose_bones['upper_arm.L'].rigify_parameters.rotation_axis = 'x'
-        metarig_pose_bones['upper_arm.R'].rigify_parameters.rotation_axis = 'x'
-
-        # fix straight leg IK problem
-        # limbs.super_limb
-        metarig_pose_bones['thigh.L'].rigify_parameters.rotation_axis = 'x'
-        metarig_pose_bones['thigh.R'].rigify_parameters.rotation_axis = 'x'
 
 
 class RigifyArmatureObject(RichArmatureObjectABC):
@@ -557,7 +325,7 @@ class RigifyArmatureObject(RichArmatureObjectABC):
         self.datapaths = datapaths
 
         self.strict_mmd_bind_infos = [
-            b for b in self.mmd_bind_infos # pylint: disable=maybe-no-member
+            b for b in self.mmd_bind_infos  # pylint: disable=maybe-no-member
             if b.pose_bone_name is not None and b.pose_bone_name in pose_bones
         ]
 
@@ -1465,7 +1233,7 @@ class MMDRigifyArmatureObject(RigifyArmatureObject):
         rig_edit_bones: bpy.types.ArmatureEditBones = self.edit_bones
         mmd_edit_bones: bpy.types.ArmatureEditBones = mmd_armature_object.strict_edit_bones
 
-        for mmd_bind_info in self.mmd_bind_infos: # pylint: disable=maybe-no-member
+        for mmd_bind_info in self.mmd_bind_infos:  # pylint: disable=maybe-no-member
             bind_bone_name = mmd_bind_info.bind_bone_name
             if bind_bone_name is None:
                 continue
