+++ mmd_uuunyaa_tools/editors/operators.py
@@ -7,7 +7,8 @@ from typing import Iterable, Union
 import bpy
 from mmd_uuunyaa_tools.editors.armatures import MMDArmatureObject
 from mmd_uuunyaa_tools.editors.armatures.autorig import AutoRigArmatureObject
-from mmd_uuunyaa_tools.editors.armatures.rigify import MetarigArmatureObject, MMDRigifyArmatureObject, RigifyArmatureObject
+from mmd_uuunyaa_tools.editors.armatures.metarig import MetarigArmatureObject
+from mmd_uuunyaa_tools.editors.armatures.rigify import MMDRigifyArmatureObject, RigifyArmatureObject
 from mmd_uuunyaa_tools.utilities import MessageException, import_mmd_tools
 
 
@@ -341,7 +342,6 @@ class MMDRigifyApplyMMDRestPose(bpy.types.Operator):
             bpy.ops.object.mode_set(mode=previous_mode)
 
         return {'FINISHED'}
-
 
 
 class MMDAutoRigConvert(bpy.types.Operator):
